rename __msvc_cxx_stdatomic.h to __msvc_cxx_stdatomic.hpp
