+++ tests/std/tests/P0943R6_stdatomic_h/test.compile.pass.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #ifndef __clang__ // TRANSITION, GH-2862
-#include <__msvc_cxx_stdatomic.h>
+#include <__msvc_cxx_stdatomic.hpp>
 
 static_assert(ATOMIC_BOOL_LOCK_FREE == 2);
 static_assert(ATOMIC_CHAR_LOCK_FREE == 2);
